requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
validators==0.22.0
Werkzeug==2.3.7
WTForms==3.0.1
